split test_sealapi in multiple files
