+++ tests/sealapi/test_context.py
@@ -1,0 +1,178 @@
+import sys, os
+import pytest
+import tenseal.sealapi as sealapi
+
+sys.path.append(os.path.dirname(os.path.abspath(__file__)))
+from utils import *
+
+
+@pytest.mark.parametrize(
+    "scheme_id,scheme",
+    [(0, sealapi.SCHEME_TYPE.NONE), (1, sealapi.SCHEME_TYPE.BFV), (2, sealapi.SCHEME_TYPE.CKKS)],
+)
+def test_encryptionparams_scheme_sanity(scheme_id, scheme):
+    assert int(scheme) == scheme_id
+
+    testcase = sealapi.EncryptionParameters(scheme)
+    assert testcase.scheme() == scheme
+
+    testcase = sealapi.EncryptionParameters(scheme_id)
+    assert testcase.scheme() == scheme
+
+
+def test_encryptionparams_scheme_specific():
+    testcase = sealapi.EncryptionParameters(sealapi.SCHEME_TYPE.NONE)
+    with pytest.raises(BaseException):
+        testcase.set_poly_modulus_degree(10)
+
+    testcase.set_poly_modulus_degree(0)
+    assert testcase.poly_modulus_degree() == 0
+    testcase.set_coeff_modulus([])
+
+    testcase = sealapi.EncryptionParameters(sealapi.SCHEME_TYPE.BFV)
+    testcase.set_plain_modulus(sealapi.SmallModulus(1023))
+    assert testcase.plain_modulus().value() == 1023
+
+    testcase.set_random_generator(sealapi.BlakePRNGFactory())
+    generator = testcase.random_generator().create()
+    assert generator.generate() != generator.generate()
+
+
+@pytest.mark.parametrize("scheme", [sealapi.SCHEME_TYPE.BFV, sealapi.SCHEME_TYPE.CKKS])
+def test_encryptionparams_scheme_settings(scheme):
+    testcase = sealapi.EncryptionParameters(scheme)
+    testcase.set_poly_modulus_degree(32768)
+    assert testcase.poly_modulus_degree() == 32768
+
+    testcase = sealapi.EncryptionParameters(scheme)
+    testcase.set_coeff_modulus([sealapi.SmallModulus(1023), sealapi.SmallModulus(234)])
+    assert len(testcase.coeff_modulus()) == 2
+    assert testcase.coeff_modulus()[0].value() == 1023
+    assert testcase.coeff_modulus()[1].value() == 234
+
+    left = sealapi.EncryptionParameters(scheme)
+    left.set_poly_modulus_degree(32768)
+    left.set_coeff_modulus([sealapi.SmallModulus(1023), sealapi.SmallModulus(234)])
+
+    right = sealapi.EncryptionParameters(scheme)
+    right.set_poly_modulus_degree(32768)
+    assert left != right
+
+    right.set_coeff_modulus([sealapi.SmallModulus(1023), sealapi.SmallModulus(234)])
+    assert left == right
+
+
+@pytest.mark.parametrize(
+    "scheme", [sealapi.SCHEME_TYPE.NONE, sealapi.SCHEME_TYPE.BFV, sealapi.SCHEME_TYPE.CKKS],
+)
+@pytest.mark.parametrize(
+    "sec_level",
+    [
+        sealapi.SEC_LEVEL_TYPE.NONE,
+        sealapi.SEC_LEVEL_TYPE.TC128,
+        sealapi.SEC_LEVEL_TYPE.TC192,
+        sealapi.SEC_LEVEL_TYPE.TC256,
+    ],
+)
+def test_context_failure(scheme, sec_level):
+    parms = sealapi.EncryptionParameters(scheme)
+
+    sealctx = sealapi.SEALContext.Create(parms, True, sec_level)
+    assert sealctx.parameters_set() is False
+
+
+@pytest.mark.parametrize(
+    "sealctx", [helper_context_bfv(), helper_context_ckks()],
+)
+def test_context_sanity(sealctx):
+    assert sealctx.parameters_set() is True
+
+
+def context_asserts(sealctx, sec_level, scheme):
+    assert sealctx.parameters_set() is True
+
+    orig_ctx_data = sealctx.key_context_data()
+    parms = orig_ctx_data.parms()
+    poly_modulus_degree = parms.poly_modulus_degree()
+    coeff_mod_count = len(parms.coeff_modulus())
+
+    def context_data_sanity(ctx_data, ctx_alias, index):
+        assert ctx_data.parms().poly_modulus_degree() == poly_modulus_degree
+        assert len(ctx_data.parms_id()) == len(ctx_alias.parms_id())
+        assert ctx_data.chain_index() == ctx_alias.chain_index()
+        assert ctx_data.chain_index() == index
+        assert (
+            ctx_data.total_coeff_modulus() != 0
+            and ctx_data.total_coeff_modulus() == ctx_alias.total_coeff_modulus()
+        )
+        assert ctx_data.total_coeff_modulus_bit_count() == ctx_alias.total_coeff_modulus_bit_count()
+        assert ctx_data.plain_upper_half_threshold() == ctx_alias.plain_upper_half_threshold()
+        assert ctx_data.coeff_div_plain_modulus() == ctx_alias.coeff_div_plain_modulus()
+        assert ctx_data.coeff_mod_plain_modulus() == ctx_alias.coeff_mod_plain_modulus()
+        assert ctx_data.upper_half_increment() == ctx_alias.upper_half_increment()
+        assert ctx_data.upper_half_threshold() == ctx_alias.upper_half_threshold()
+
+        qualifiers = ctx_data.qualifiers()
+        assert qualifiers.parameters_set is True
+        assert qualifiers.using_fft is True
+        assert qualifiers.using_ntt is True
+        assert qualifiers.using_batching is True
+        assert qualifiers.using_fast_plain_lift == (scheme == sealapi.SCHEME_TYPE.BFV)
+        assert qualifiers.sec_level == sec_level
+
+    for (parms_id, ctx_data, index) in [
+        (sealctx.first_parms_id(), sealctx.first_context_data(), coeff_mod_count - 2),
+        (sealctx.last_parms_id(), sealctx.last_context_data(), 0),
+        (sealctx.key_parms_id(), sealctx.key_context_data(), coeff_mod_count - 1),
+    ]:
+        should_be_same_ctx = sealctx.get_context_data(parms_id)
+        context_data_sanity(ctx_data, should_be_same_ctx, index)
+
+    assert sealctx.last_context_data().next_context_data() is None
+    assert (
+        sealctx.first_context_data().prev_context_data().chain_index()
+        == sealctx.key_context_data().chain_index()
+    )
+
+    assert sealctx.using_keyswitching() is True
+
+
+@pytest.mark.parametrize(
+    "sec_level",
+    [sealapi.SEC_LEVEL_TYPE.TC128, sealapi.SEC_LEVEL_TYPE.TC192, sealapi.SEC_LEVEL_TYPE.TC256,],
+)
+def test_context_scheme_bfv_sanity(sec_level):
+    poly_modulus_degree = 8192
+    plaintext_modulus = 1032193
+
+    parms = sealapi.EncryptionParameters(sealapi.SCHEME_TYPE.BFV)
+    parms.set_poly_modulus_degree(poly_modulus_degree)
+    parms.set_plain_modulus(plaintext_modulus)
+
+    coeff = sealapi.CoeffModulus.Create(poly_modulus_degree, [32, 32])
+    parms.set_coeff_modulus(coeff)
+    sealctx = sealapi.SEALContext.Create(parms, True, sec_level)
+    assert sealctx.parameters_set() is True
+
+    coeff = sealapi.CoeffModulus.BFVDefault(poly_modulus_degree, sec_level)
+    parms.set_coeff_modulus(coeff)
+    sealctx = sealapi.SEALContext.Create(parms, True, sec_level)
+    context_asserts(sealctx, sec_level, sealapi.SCHEME_TYPE.BFV)
+
+
+@pytest.mark.parametrize(
+    "sec_level", [sealapi.SEC_LEVEL_TYPE.TC128],
+)
+def test_context_scheme_ckks_sanity(sec_level):
+    poly_modulus_degree = 8192
+    plaintext_modulus = 1032193
+
+    parms = sealapi.EncryptionParameters(sealapi.SCHEME_TYPE.CKKS)
+    parms.set_poly_modulus_degree(poly_modulus_degree)
+
+    coeff = sealapi.CoeffModulus.Create(poly_modulus_degree, [60, 40, 40, 60])
+    parms.set_coeff_modulus(coeff)
+    sealctx = sealapi.SEALContext.Create(parms, True, sealapi.SEC_LEVEL_TYPE.TC128)
+    assert sealctx.parameters_set() is True
+
+    context_asserts(sealctx, sec_level, sealapi.SCHEME_TYPE.CKKS)
